requirements_dev.txt
amqp==5.1.1
anyio==3.7.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==41.0.1
fastapi==0.96.0
grpcio==1.54.2
grpcio-tools==1.54.2
h11==0.14.0
hiredis==2.2.3
idna==3.4
kombu==5.3.0
more-itertools==8.14.0
netaddr==0.8.0
pptree==3.1
prompt-toolkit==3.0.38
protobuf==4.23.2
pycparser==2.21
pydantic==1.10.8
python-dotenv==1.0.0
python-ulid==1.1.0
pytz==2023.3
redis==4.5.5
redis-om==0.1.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
types-pyOpenSSL==23.2.0.0
types-redis==4.5.5.2
typing_extensions==4.6.3
urllib3==2.0.2
uvicorn==0.22.0
vine==5.0.0
wcwidth==0.2.6
